Refactor object pools
